allow number of addresses to return -1 to indicate that the search just started for command line interface

+++ src/kad.h
@@ -21,7 +21,8 @@ int kad_blacklist(const IP* addr);
 * Lookup the addresses of the nodes who have announced value id.
 * The first call will start the search.
 */
-int kad_lookup(const char query[], IP addr_array[], size_t *addr_num);
+
+int kad_lookup(const char query[], IP addr_array[], int *addr_num);
 
 // Export good nodes
 int kad_export_nodes(FILE *fp);
